Remove api-extractor/api-documenter (#6528)
* Remove api-extractor/api-documenter

api-extractor with api-documenter allows generation of Markdown files from TS APIs. We run it for react-native-windows and react-native-win32, created some automated markdown files.

These files don't escape our repo, and are fairly low quality in terms of documentation. We're removing or discouraging new TS for APIs as well, leading to some of this disappearing over time, or becoming less complete. It's not likely these have a future, but we see frequent package updates from these, have to lint for them, and need to generate a number of unused files because of them.

This changes removes usage of api-documenter/api-extractor, the corresponding tasks, and the generated docs from them.

* Change files

* Remove yarn api step