Pack Win32 NuGet Package in PR
This was removed when inlining NuGet packing for MSRN into init tests. It was thought at the time the loss of coverage would be low enough not to cause issues, but we've been seeing issues pop up nonetheless.

Wer add a job which on desktop PR completion, reads artifacts and pumps out a desktop specific NuGet package. This adds some time to the desktop critical path, but this will be mitigated by removal of integration tests in the future. We also remove some desktop slices, after asks to reduce flavors. We don't do the NuGet packing inline with desktop PRs because it requires a little bit more PR work, and we should be able to take a bit less PR time by doing the packing in a single job. Unline experimental NuGet tests, we don't need to bind to npm version which is inconsitent per-job, which lets us have it be fully independent like before.