Layout-Headers During NuGet Pack
PR and publish are set up to expect that some NuGet processing work is done before uploading artifacts. It is expected a single job does this work, and uploads for all NuGet consumers. If multiple jobs due this work, we would potentially see conflicting artifacts.

This work also uploads build scripts as an artifact, leading to a race condition where the Desktop Packing job might fail if Desktop finishes before the designated job to upload header artifacts.

These headers can now be fully derived from source in the repo, so we can remove some of this fragility and group the work with NuGet packing.