Wait until UWP Jobs are Finished Before Desktop Packing
NuGet generation in CI is dependent on a single job creating and uploading shared headers for all NuGet packages. This job requires doing the build, but seemingly cannot be done by multiple machines, as they would upload overlapping artifacts.

This is currently done in PR by one of the UWP build steps. Ensure these are completed before packing the desktop NuGet package to avoid the race condition.

We should consider splitting out desktop and UWP NuGet generation artifacts in the future to avoid tying them together in this way.