Update lib template to default prefer RNW dir of consuming app (#6577)
This update to the lib templates helps fix the issue where native module repos also have example apps in their codebase, and two copies of M.RN get restored and built against.

When setting for ReactNativeWindowsDir, we start searching for RNW based on the location of the project file and looking upwards for node_modules/react-native-windows. In the case of a regular app repo with a single root package.json, where every native module is under the app's node_modules, the app and native module projects all find the same version.

In other repo layouts, we run into cases with multiple package.json files, and where the app is lower in the folder hierarchy than the native module it's consuming (such as an example app within a native module repo, as per issue #6576). You can then up with multiple copies of RNW, and worse, the app can build against one version while the native module builds against another, and then build/packaging fails.

However, since we always build from a single solution file, by instead starting the search for ReactNativeWindowsDir at the location of the solution file, we can make sure that all the modules build against the copy of RNW that the app building against.

Closes #6576