Use strings for telem exit codes
Fixes #6610

I missed during eslint rule enabling that we were converting user exit codes from integers back to the original strings when sending to telem. This change makes us send the string instead of int for telem to keep schema consistency (although a larger telem payload).

Instead of the enum method of object access to integer, then reverse mapping from ordinal back to string, we just pass the string as a union type and convert to int from the map in the place where we set process exit code.