Fix: monorepo page not found
When entering the 'building-rnw.md' page, Monorepo is mentioned as a
type of RNW. The issue is that when clicking on the "this page" link,
404 is returned instead of redirecting to the monorepo.md page.
The root cause of that issue is the capital letter originally used in
the link code, while the page itself has a name as: "monorepo.md".

This commit provides the fix for that by changing the letter to the
correct one, so there's no confusion for newcommers.