Enhance the generateCertificate error result
The `generateCertificate()` function had the try-catch result given only
as the general failure message.
To keep the user aware of issues during the RNW setup, this self-signed
certificate generating try-catch block has been provided with the
additional information included in the exception argument: err.

Please note, that only the err is displayed, instead of extending it
with additional parameters (such as ERRNO, or path) as these are
developers details and it keeps the error message informative but still
short.