Remove the timeout from self-signed certificate generating
The timeout set to 10 seconds when generating the self-signed
certificate was causing the error on slower machines for developers.
Please see:
https://github.com/microsoft/react-native-windows/issues/6281

The idea was also to make this timeout value configurable, but in the
end this solution would just make each developer leave it as default
(magic number 10) or set it to max just so it's not needed to be set
again. This could also result in trials and errors when trying to set
the proper timeout value, which would make using the RNW uncomfortable.