Ensure windows script added on init
Fixes #5931

At some point we started only re-installing RN dependencies in generator if the package peer dependencies were unsatisfied. This missed some logic grouped in to also modify package.json to add a windows script.

This change fixes that, and does some refactoring to use package utils, and not create messages that w're doing a yarn install when we don't in the module project unmet peer dependency case.

Still need to validate this