Allow unreleased RNW versions with react-native-windows-init --useDevMode
When cutting an unreleased version we may not have an available npm package. We can link a local unreleased package, but doing a yarn install will complain if the version is not found in the npm registry. This change modified useDevMode logic to instead specify a wildcard version, which keeps yarn happy with the linked but unreleased version.