[0.63] Backporting TimingModule assert removals (#6653)
* Remove Sporadic Assert on Instance Reload from TimingModule (#5988)

* Remove Sporaddic Assert on Instance Reload from TimingModule

TimingModule will sometimes assert when reloading the instance. Timing::OnRendering is called in a callback off-thread, so it seems like during instance reload our message to delivery a timer event to JS happens after the instance has already been destroyed while the module ahsn't been quite yet.

Did a quick scan for any memory safety/ref cycle issues in case we were keeping this module alive too long, but didn't find anything. This change removes the assertion, since it should be harmless.

When looking for potential memory issues, I also discovered we expose module creation to the Win32 ABI for some reason even though this isn't called from devmain, so that is removed.

* Change files

* Revert removed exports

* Remove the other TimingModule Assertion (#6174)

* Remove the other TimingModule Assertion

TimingModule methods are dispatched to the UI thread. The instance/Module may be gone by the time its methods are processed in the UI queue. We removed an assertion that an instance is available on render callback, but there is a second assertion on method call as well. Remove the assertion.

* Change files

* change type from "prerelease" to "patch"

Co-authored-by: Nick Gerleman <ngerlem@microsoft.com>