Fix dependency syncing for wildcard peer dependencies (#6659)
* Fix dependency syncing for wildcard peer dependencies

As part of integrating a new release of react-native we run an algorithm to sycnrhonize dependencies, peer dependencies, and devDependencies across the repo. Part of this involves updating any peer dependencies to react-native that would be broken by a new version. We do that only if the existing peer dependency is unsatisfied by the new version.

The semver package has some suprising behavior when testing satisfaction of wildcard peer dependencies, where it does not count any prerelease build as satisfying it. This leads to us to try to bump the wildcard semver and fail if we encounter it.

Special case wildcard when doing semver satisfaction checks.

* Change files