Cover winrt::hstring WriteValue with unit tests
Unit tests for the hstring type of value checks following cases:
 - if value written as default wchar_t type is interpreted correctly and
   there's no confusion between other wchar types,
 - if value written as winrt::hstring type is interpreted as wchar type
   which should result in the same characters checked as output,
 - if value written as winrt::param::hstring is covered separately as a
   correct type which isn't confused with any other wchar type and the
   output characters result matches the input.

NOTE: Explicitly specified types for WriteProperty are not required in
the real user scenario. They are used in the unit tests to ensure that
this exact given type of string will be passed to the WriteValue so
there's a check whether this exact type is parsed and written correctly.