Reader/Writer for winrt::hstring (#6704)
* Implement WriteValue for hstring

This implementation adds the WriteValue functions overloads for both:
winrt::hstring and winrt::param::hstring
values.
Adding ::param usage covers the subset of winrt::hstring if it's used
explicitly - without it, for some compilers, it would still run into the
is_convertible condition for templated WriteValue implementation.

* Implement ReadValue for hstring

The implementation of ReadValue for the winrt::hstring is based on other
implementation providing support for std::string and std::wstring.
The differences are:
 - to_hstring used for conversion, as this reduces the inefficiency in
   types matching
 - double type converted directly without erasing the potential extra
   zeros or dots.
 - winrt::param::hstring is not supported for ReadValue as this type
   should not be used directly.
   Moreover: implementing support for winrt::param::hstring would not
   introduce any benefits, as it's implementation would be exactly the
   same as the basic winrt::hstring.

* Change files

* Run yarn format on implementation

* Cover review findings

Changes marked by the reviewers are:
 - Keep the value manipulation when converting double type by the
   winrt::hstring reader.
   This manipulation keeps the ValueReader results consistent for each
   string type.
 - Make the value for winrt::hstring writer a const reference instead of
   const values.
 - Cover the implementation with unit tests according to other string
   types testing procedures.

* Cover winrt::hstring WriteValue with unit tests

Unit tests for the hstring type of value checks following cases:
 - if value written as default wchar_t type is interpreted correctly and
   there's no confusion between other wchar types,
 - if value written as winrt::hstring type is interpreted as wchar type
   which should result in the same characters checked as output,
 - if value written as winrt::param::hstring is covered separately as a
   correct type which isn't confused with any other wchar type and the
   output characters result matches the input.

NOTE: Explicitly specified types for WriteProperty are not required in
the real user scenario. They are used in the unit tests to ensure that
this exact given type of string will be passed to the WriteValue so
there's a check whether this exact type is parsed and written correctly.