Use CallJSFunction for DeviceInfoModule
DeviceInfoModule is responsible for pulling the dimensions of both
window and screen of the application. It uses the EmitJSEvent to send
the gathered data from native side to JS side.
After applying the fix for window's dimensions always at fixed values
(please see: https://github.com/microsoft/react-native-windows/pull/6286)
the DeviceInfoModule would not work unless dependency fix is applied.

This fix has been noticed as a breaking change for 0.63 version so this
commit provides the fix with a workaround:
As CallJSFunction is already wrapped by the EmitJSEvent method it is now
used instead, so the explicitly specified "emit" event is correctly
launched.