Allow Integration Script to Output Markdown
This adds a flag to the RN integration script to generate a markdown report with status of checks. This will be used for automated integration as the body of pull requests created by the bot.

We additionally automate the addition of a URL in reports to show new commits since last integration.

Added some simple UTs for the logger and manually validated against an integration.