Use octokit instead of raw GitHub REST APIs (#6879)
We're going to be adding more code that uses the GitHub API. This increase in usage makes it worth using GitHub's official API library instead of rolling our own REST code. This allows some benefits such as built-in typing, automatic errors, plugins to retry on failure, etc. Use this in the release creation script instead of using node-fetch directly.

The dependency is heavy enough to not make sense in this isolated use-case, but makes more sense with icreases usage in future repo tooling.