Stand up Bot Coordinator Azure Function
This change adds a new package for the bot coordinator, adapting a template function app to engineering systems in our repo. Right now it does very little, and it mostly just structure and deployment.

Deployment of the function app is tied to nightly publish, partially for ease of use, but also to ensure changes to monorepo packages are reflected.

Validated building/running the function app locally. Have tested bits of the deploy, but not everything end to end. That might need some iteration based on how Azure Pipelines likes it, but its job is independent of the rest of publish so it shouldn't degrade it (apart from potential YAML parse errors).