Try to fix bot deployment
Pipeline logic in bot deployment wasn't quite correct, which isn't super unexpected. This change fixes a couple of issues before trying to test deployement again.

1. Setting variables for the yarn install template only in the case of frozenLockfile: true, can create ordering issues
2. Powershell scripts with pipe characters on the wrong line

We also declare the bot coordinator as a public package so that beachball will warn if it depends on a private package (as those cannot work outside the monorepo).