Avoid crash when unsupported facade type is given
Starting the animation for unsupported facade type led to a crash of the
application.
This was caused by the assertion which was checking the given type
against the last of supported types in RNW. So to prevent it from
crashing,  `None` type has been introduced.

This new type's responsibility is to indicate the unsupported type
received and to pass this notification to the methods calling this
conversion.