TypeScript Stack Traces and Zero Build for CLI Apps
Playing around in a few packages with the idea of using "ts-node" to do TS transpilation dynamically during node requires.

This has some clear benefits like not needing to build for package changes, and getting correct line numbers in stack traces during development. It unfortunately also adds a ~3 second startup to the CLI (though we seem to be loading some extra modules we could optimize).

It runs using tsc and tsconfig, so there should be low risk of different code generation. Want to see if this runs into any issues or ends up being useful before applying this pattern to other packages.