Move office-iss packages to scope
Fixes #6937

The majority of packages in `packages` directory follow their package name, but @office-iss/react-native-win32 does not. Move the directory into a folder for scoped package, and move react-native-win32-tester along with it.

This also removes some stray gitignored files from react-native-wi32, so we clean up these outputs in vnext as well.