Node to UWP RPC Package (#7071)
* Node to UWP RPC Package

This change implements a package to allow Node applications to make generic RPC calls to RNW applications. To bypass UWP restrictions, the implementation has the Node application create a server which the RNW app polls for to initiate connection. Installation is through autolinking.

This enables scenarios for test apps, such as returning tree-dumps to the E2ETest test runner through the wire. To begin with, we just replace the pipe used by integration-test-app.

More information in the README added.

Note that I haven't yet validated the change, and there are still some TODOs. I plan to address these before merge, and do not expect structure/implementation to change too drastically.

* yarn format

* Remove inbound loopback exemption

* Working enough to pass tests

* build fix

* Bug numbers and typo fixes

* Fix version