Fixes missing image size in Image onLoad event
As reported in #7172, the native event payload does not contain the
pixel width and height of the image in standard cases (that use
BitmapImage). The issue was that the metadata for width and height sent
in the event was not updated until the XAML ImageSource.ImageOpened
event fired, and the onLoad event itself was triggered by the
BitmapImage.ImageOpened event. The BitmapImage.ImageOpened event will
always fire before it's parent ImageSource.ImageOpened event, thus the
event emitted to React Native would never have up-to-date information on
the image size.

This commit moves the logic to update the image size metadata from the
ImageSource.ImageOpened event to the BitmapImage.ImageOpened event.

Fixes #7172