Include Hermes NuGet Package By Default
Note that this is waiting on a new Hermes package 0.7.2-microsoft.3 after a bug in targets discovered in 0.7.2-microsoft.2. This should be close to the final change though.

This change leads to the inclusion of the Hermes NuGet package by deault. It being copied to output is conditioned on UseHermes being enabled in Microsoft.ReactNative.vcxproj.

It also fixes Hermes builds in general, which seem to have broken, and where Hermes CI jobs were not passing correctl flags.

This does not yet enable conditional compilation of Hermes branches unless flag is turned on. We do enable Hermes by default in Playground when web debugging is disabled, since Chakra stack is more mature, and eyes on Hermes could be useful. I changed Playground UI to gray out engine selection if web debugging is checked to make clear we are not using Hermes if checked.

NuGet deps: +7.57 MB
App package size (default): +0MB
App package size(opt in release x64): +2MB

Binaries not available for ARM32. Still need to update docs after this change.

Manually tested Playground with Hermes flags set in solution. Some coverage of the scenario also exists.