Coalesce ScrollView Messages
Fixes #7237
Fixes #7256

The overall goal of this change is to fix performance issues impacting FlatList. See the above issues for more details.

This is largely based on what Android does, where they wait until the end of a frame, then queue a message to JS to pump the batch (with coalescing).

This is exposed to the ABI, but is on the periphery instead of being added to ReactContext.