Narrow fix for tintColor and repeat images (#7222)
* Narrow fix for tintColor and repeat images

There is a [bug](https://github.com/microsoft/Win2D/issues/809) in
GaussianBlurEffect from Win2D that causes images to disappear when
moving the app between screens. Since the ReactImageBrush always created
the GaussianBlurEffect (even when `blurRadius` was 0) whenever
`resizeMode=repeat` or `tintColor` was used on the Image, this meant
that Images that only used `resizeMode` or `tintColor` suffered from the
same bug.

This change ensures that GaussianBlurEffect is only used when the
`blurRadius` prop is non-zero.

Fixes #7205

* Change files