Coalesce ScrollView Messages (#7260)
* Coalesce ScrollView Messages

Fixes #7237
Fixes #7256

The overall goal of this change is to fix performance issues impacting FlatList. See the above issues for more details.

This is largely based on what Android does, where they wait until the end of a frame, then queue a message to JS to pump the batch (with coalescing).

This is exposed to the ABI, but is on the periphery instead of being added to ReactContext.

* Change files

* whoops

* - Remove from ABI, so we can iterate on interface privately
- Move to emitter-per-viewmanager instead of emitter-per-instance. This helps with removing conflicts and keeping short queue length
- Remove coalesceKey in favor of automatically partitioning based on event name + tag
- Make friendlier to C++ non/ABI generated types
- Lots of renaming and documentation to make intent clearer
- A couple of internal changes

* More fixes

* PR feedback

* yarn format

* Fix comment typos