Change TextInput cursor behavior (#7285)
* Change TextInput cursor behavior

As stated in issue #7283, the cursor position changes seem to be
somewhat different on each platform. Since we're already adding code in
react-native-windows to change the Windows default cursor positioning
behavior, we might as well make it a bit more consistent with other
platforms. The change here matches the behavior on iOS, macOS and Web
(so it's only different from Android, which seems to keep cursor
position relative to the TextInput value length constant).

Fixes #7283

* Change files