Fixes issue with Run text getting overwritten
There are a few issues with how `textTransform` is currently
implemented, but one particular issue is that the shallow application
algorithm for the `textTransform` inadvertently applies the text of the
last Run in the InlineCollection to all runs in the collection when
trying to apply an inherited transform.

Please note, this fix is a temporary patch to eliminate the specific bug
described in #7534. There is a more comprehensive fix that I will send a
follow-up PR for related to #7535.

Fixes #7534