Fixes textTransform recursion
This change includes a number of fixes to the approach to textTransform.
There were 4 issues with the previous textTransform approach that this
diff solves:
1. The textTransform prop was only applied to the parent TextBlock when
it had a single raw text child
2. The textTransform prop on a virtual text Span only recursed one level
3. The one-level textTransform approach had a bug where the last raw
text value overwrote all Run text values (#7534)
4. The textTransform prop did not update correctly when changed

Rather than holding state in the TransformableText struct, this change
keeps the originalText state exclusively in RawTextShadowNode, and only
when a parent VirtualTextShadowNode or TextShadowNode has a
textTransform setting (that is not "none" or undefined). The
VirtualTextShadowNode and TextShadowNode holds only the TextTransform
prop value (and not any originalText value which was causing the issue
in #7534). When a child is added to either a TextShadowNode or
VirtualTextShadowNode, we now attempt to recursively apply the
textTransform on the child sub-tree (wherever textTransform is not
explicitly overwritten). Whenever the textTransform prop changes on
VirtualTextShadowNode or TextShadowNode, we now attempt to recurse over
all children of that node to apply the textTransform (wherever it is not
explicitly overwritten).

Fixes #7534
Fixes #7535