Fixes bug where raw text update overwrites all text
After fixing issues with textTransform, 278a3c13 introduced a bug where
updates to a raw text node can overwrite the TextBlock parent if it has
only a single Span child inline. This change ensures we check that the
updated raw text Run is indeed the direct child of a TextBlock.