Bump xmldom to ^0.5.0. (#7516)
There is a security vulnerability in xmldom versions below 0.5.0. This PR adds a yarn resolution to force xmldom to be at 0.5.0.

The dependency chain that brings in the vulnerable xmldom is as follows: react-native@0.0.0-d477f8011 -> @react-native-community/cli-platform-ios@^5.0.1-alpha.0 -> xcode@^2.0.0 -> simple-plist@^1.0.0 -> plist@^3.0.1 -> xmldom "0.1.x. So bumping the xmldom used here would require considerable work, which is why we use yarn resolution here.

Backport of #7514.