Moved Component Governance task into template (#7686)
This PR consolidates some of the logic for running the Component Governance task into `templates\component-governance.yml`.

Specifically, the logic that, if EnableCodesign == true, CG must fail the job if there are any alerts. FYI, in our whole ecosystem of pipelines, EnableCodesign should only be true in the Publish pipeline against stable branches.

This PR also adds a CG check _before_ beachball bump in Publish, so the publish can be aborted without publishing to NPM and GitHub.