Allow `lineHeight` settings less than font height
Since the default LineStackingStrategy for XAML TextBlock uses
`LineStackingStrategy::MaxHeight`, when the <Text> `lineHeight` prop is
set to a value smaller than the calculated max height for the text, it
is ignored.

Setting LineStackingStrategy to `LineStackingStrategy::BlockLineHeight`
enables `lineHeight` settings smaller than the font height, and has
identical behavior when the `lineHeight` value is larger than the max
font height.

Fixes #7695