Emit TextInput onChangeText before onSelectionChange (#7742)
* Emit TextInput onChangeText before onSelectionChange

Previously, TextInputViewManager subscribed to two events to implement
`onChangeText`:
1. TextChanging - used to increment a native counter to ensure JS
changes to text do not overwrite native changes to text.
2. TextChanged - used to emit the `onChangeText` event and surprisingly
also incremented the native counter.

Additionally, the view manager subscribed to SelectionChanged to emit
the `onSelectionChange` event.

In XAML, the sequence of native events is as follows:
1. TextChanging
2. SelectionChanged
3. TextChanged

Since we emit the event that ultimately triggers `onChangeText` via the
TextChanged native event, this means that we actually receive the cursor
position change before the text change. This is different from other
React Native platforms and makes it more difficult to implement
behaviors based off the se`onSelectionChange` event.

One concern with this change is if there are any repercussions of
eliminating the two phased approach to the native counter and firing of
the event. Here is why I don't believe this will create any issues:

1. Previously, we incremented `m_nativeEventCount` twice, once in
`TextChanging` and once in `TextChanged`. Reducing this to incrementing
only once will not introduce any bugs, as JavaScript does not increment
the `mostRecentEventCount` prop, so the only requirement is that a
native change to the TextBox content increases the count.

2. Previously, we emitted the event in `TextChanged`. It appears that
the `TextBox.Text()` property is already updated at the time that
`TextChanging` is called, whether the `Text()` property was set
externally or it was updated internally by a key event. The only
difference between the two events is that `TextChanged` is called after
the text is rendered (which may occur asynchronously from the actual
text change). This is the reason we needed to introduce the dependency
on `TextChanging` in the first place, as `TextChanging` is guaranteed to
be called synchronously with the value change (and could not be
interrupted by, e.g., an update from JS).

3. Is it a problem that JavaScript receives the text changed event
before it actually renders? This is a good question, but I can't think
of any scenarios where it's important that the TextBox has actually
rendered the change before emitting the event, and since JS runs
asynchronously from the UI thread, it's likely that by the time any UI
changes / side effects triggered by JS make their way back to the UI
thread, the text will have been rendered.

Fixes #7740

* Change files

* Applying yarn format

* Adds E2E test for order of events

* Adding back inadvertently removed guard

* yarn lint:fix