Allow pointer events on selectable Text
When TextBlock::IsTextSelectionEnabled is set to true, pointer events
are canceled because the events are always marked as handled by the
TextBlock, thus the events never reach the React Native root view. This
change subscribes a TouchEventHandler to Text when the `selectable` prop
is set to true, and also conditionally fires the `onTouchEnd` event when
PointerCaptureLost fires when text is *not* selected as a result of the
pointer events (e.g., click-and-drag or the second click of a double
click). This allows us to create `onPress` callbacks that mimic the
behavior of the XAML hyperlink component. In order for the correct
page layout dimensions to be reported, we also need to provide the root
view to TouchEventHandler.

Fixes #7812