Use Rex With V8 JSI Runtime (#7681)
* [Do not merge yet] Use Rex With V8 JSI Runtime

Fixes #5807

A new version of Rex will default to using a JSI runtime with V8. This will let us remove some old JS patching to enable 0.64 usage with previous native ChakraExecutor. The version of Rex with the change is not yet available, so this PR currently causes run-win32 to crash on startup.

* Change files

* Use V8 for real now