Run integration-test-app against both Chakra and Hermes
While we have Hermes build validation, we do not yet have runtime tests against Hermes. e2e-test-app would be a good candidate, but is C#, which Hermes does not yet support. Instead I updated integration-test-app to add `ExperimentalFeatures.props`, and do add a script to the CI matrix which enables Hermes in it, then runs in release (for no web debugging).

This silently conflicts with https://github.com/microsoft/react-native-windows/pull/7802 which updates to a newer version of Hermes than added here.