Run integration-test-app against both Chakra and Hermes (#7806)
* Run integration-test-app against both Chakra and Hermes

While we have Hermes build validation, we do not yet have runtime tests against Hermes. e2e-test-app would be a good candidate, but is C#, which Hermes does not yet support. Instead I updated integration-test-app to add `ExperimentalFeatures.props`, and do add a script to the CI matrix which enables Hermes in it, then runs in release (for no web debugging).

This silently conflicts with https://github.com/microsoft/react-native-windows/pull/7802 which updates to a newer version of Hermes than added here.

* Add matrix to job name

* Wrong order

* Manually set BundleEntryFile

* Use faster pool to avoid lengthening critical path

* Enable AppX Signing

* Remove duplicated comment

* Disable failing tests

* Avoid temp cert

* Change --no-deploy pipeline logic

* Change files

* fix typo

* Add --deploy-from-layout to both PR and CI

* key artifacts to matrix name

* Remove GenerateAppxPackageOnBuild

* Try reverting to builtin pool

* Rev Hermes

* Add option to CLI README

* Proper XML manipulation