Adds event coalescing to TouchEventHandler
This change adds event coalescing to onTouchMove events. The previous
implementation of BatchingEventEmitter expected a view tag and event
name to perform coalescing. This change generalizes the batching event
emitter to accept a coalescing key instead, which simply concatenates
the event name and the view tag in the case of ScrollViewViewManager and
the event name and the pointer ID in the case of TouchEventHandler. This
change also modified the signature of Emit(Coalescing)JsEvent to accept
a JSValueArgWriter rather than a JSValue, as the arguments sent to the
`RCTEventEmitter.receiveTouches` JS function do not exactly match the
arguments sent to `RCTEventEmitter.receiveEvent`.

Fixes #7276