Patch Community CLI Metro in RNW CLI postinstall
For 0.65 I am wanting to experiment in our preview with using an RN release decoupled from the upstream stable branch. We do not need most 0.65 upstream cherry-picks so far, but we do need a CLI change to use a newer version of Metro. This fixes some user-facing bugs including bad errors and fast refresh being broken.

To accomplish this I make our CLI pull in the correct versions of Metro, then patch over the upstream CLI.

Will create an issue to revert this if/when we decide to use a non-nightly build.

Validated a "yarn install" in our monorepo will apply the patches. Not making this in our master branch since we will integrate a new version with the fix.