Emit `onSubmitEditing` before clearing text (#7949)
* Emit `onSubmitEditing` before clearing text

When creating logic for submitting editing, apps may wish to use the
current state of the text set by the `onChange` event rather than the
text passed through by the `onSubmitEditing` event, especially if it's a
controlled component where JS may modify the text in the `onChange`
event. The recent change to emit `onChange` events in the
`TextBox.TextChanging` event rather than the `TextBox.TextChanged` event
cause the `onChange` event to fire synchronously when the `TextBox.Text`
value is cleared, and the `onChange` event was arriving before
`onSubmitEditing`.

* Change files