[0.65] Add experimentalFeatures to projectConfig result
This PR backports #8402 (also #7980 and #7982) to 0.65.

Original commit messages:

Add experimentalFeatures to projectConfig result (#8402))

Experimental, build-wide settings (such as using the RNW nuget packages, or using Hermes instead of Chakra, or using WinUI3) are set in the `ExperimentalFeatures.props` file in the solution directory of RNW apps.

Sometimes we want the RNW CLI itself to behave differently given this information, specifically we want to report some of this data via telemetry.

This PR:

* Adds parsing of the properties in this props file to the project config (i.e. the windows result from calling `react-native config`)
* Moves the `UseExperimentalNuget` flag out of the project templates and into the `ExperimentalFeatures.props` file, so that modules loaded from elsewhere (say node_modules) can also access the property's value (keeping an app and it's modules in sync)
* Adds logic to make sure that the config still detects `UseExperimentalNuget` present within the project files of projects built using the old template logic
* Adds new tests to cover this new functionality
* Cleans up the project config tests to remove the unused "expected" objects when the the jest snapshots are doing the real validation

Disable project/dependency config when not running on windows (#7980)

This PR adds a check to project and dependency config to bail when not
running on windows, since we assume non-windows machines don't need the
windows config information.

Normalize paths provided by react-native.config.js (#7982)

* Normalize paths provided by react-native.config.js

This PR makes sure that raw paths specified in a react-native.config.js
are normalized when config is run, before the result is passed up to
whichever CLI command called for it.

Closes #7979