Use DispatcherTimer unless animation frame needed
The CompositionTarget::Rendering event burns a lot of CPU. For long
running interval timers that fire infrequently, using this event for the
timer implementation is overkill. This change checks the period of the
timer, and uses the CompositionTarget::Rendering callback for period ==
1ms (which is typical of requestAnimationFrame requests) and for
everything else, uses a dispatcher timer that updates it's tick to the
next due time each time a timer fires.