Tie visitAllPages Test to RNTesterList (#8071)
* Tie visitAllPages Test to RNTesterList

Fixes #7726

We have a UIA-based test which navigates back and forward from each RNTester example. The list of examples cannot be populated by the UI, and is currently hardcoded. The source of truth, RNTesterList in RNTester, loads quite a bit of RN with it.

A previously discussed approach was to relay loaded data from the RN application process to the test runner. This change goes with a simpler approach, of making the test runner act more like RN. That includes Metro's Babel transformations, Windows platform resolution logic, and global setup. We can mostly rely on existing react-native UT infrastructure to automate most of this.

This loads enough of the environment to allow successfully loading all RNTester example modules, letting us poke at its data.

* remove extraneous export

* Reduce synced typings surface